Add SortOrderManager and SortOrderVariety interface docs
